requirements.txt
isodate==0.6.0
google-api-python-client==2.14.0
google-auth-oauthlib==0.4.4
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.0
python-dateutil==2.8.2
